TST: reorganize misplaced quarter / business quarter tests (#27085) (#42932)
